fix ui colors/shapes for v2

+++ src/main/resources/com/bpodgursky/hubris/db/migration/v3_create_game_syncs.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS game_syncs(
   id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
   cookies_id INT NOT NULL,
-  game_id INT NOT NULL,
+  game_id BIGINT(20) NOT NULL,
   UNIQUE(cookies_id, game_id),
   FOREIGN KEY (cookies_id)
     REFERENCES np_cookies(id)
